chore: Refactor generateFlagUrl function for better code organization and readability
